+++ rollup.config.mjs
@@ -1,11 +1,12 @@
 import commonjs from '@rollup/plugin-commonjs';
 import { nodeResolve } from '@rollup/plugin-node-resolve';
+import terser from '@rollup/plugin-terser';
 import typescript from '@rollup/plugin-typescript';
 import { readFileSync } from 'fs';
-import { precompile } from 'handlebars';
+import handlebars from 'handlebars';
 import { dirname, extname, resolve } from 'path';
-import externals from 'rollup-plugin-node-externals';
-import { terser } from 'rollup-plugin-terser';
+
+const { precompile } = handlebars;
 
 /**
  * Custom plugin to parse handlebar imports and precompile
@@ -28,6 +29,7 @@ const handlebarsPlugin = () => ({
                 preventIndent: true,
                 knownHelpersOnly: true,
                 knownHelpers: {
+                    ifdef: true,
                     equals: true,
                     notEquals: true,
                     containsSpaces: true,
@@ -47,9 +49,6 @@ const handlebarsPlugin = () => ({
 
 const getPlugins = () => {
     const plugins = [
-        externals({
-            deps: true,
-        }),
         nodeResolve(),
         commonjs({
             sourceMap: false,
@@ -72,5 +71,6 @@ export default {
         file: './dist/index.js',
         format: 'cjs',
     },
+    external: ['camelcase', 'commander', 'fs-extra', 'handlebars', '@apidevtools/json-schema-ref-parser'],
     plugins: getPlugins(),
 };