+++ src/openApi/v3/parser/getModel.ts
@@ -23,6 +23,7 @@ export const getModel = (
         template: null,
         link: null,
         description: definition.description || null,
+        deprecated: definition.deprecated === true,
         isDefinition,
         isReadOnly: definition.readOnly === true,
         isNullable: definition.nullable === true,
@@ -94,9 +95,13 @@ export const getModel = (
         }
     }
 
-    if (definition.type === 'object' && typeof definition.additionalProperties === 'object') {
-        if (definition.additionalProperties.$ref) {
-            const additionalProperties = getType(definition.additionalProperties.$ref);
+    if (
+        definition.type === 'object' &&
+        (typeof definition.additionalProperties === 'object' || definition.additionalProperties === true)
+    ) {
+        const ap = typeof definition.additionalProperties === 'object' ? definition.additionalProperties : {};
+        if (ap.$ref) {
+            const additionalProperties = getType(ap.$ref);
             model.export = 'dictionary';
             model.type = additionalProperties.type;
             model.base = additionalProperties.base;
@@ -105,7 +110,7 @@ export const getModel = (
             model.default = getModelDefault(definition, model);
             return model;
         } else {
-            const additionalProperties = getModel(openApi, definition.additionalProperties);
+            const additionalProperties = getModel(openApi, ap);
             model.export = 'dictionary';
             model.type = additionalProperties.type;
             model.base = additionalProperties.base;
@@ -145,12 +150,12 @@ export const getModel = (
     }
 
     if (definition.type === 'object') {
-        model.export = 'interface';
-        model.type = 'any';
-        model.base = 'any';
-        model.default = getModelDefault(definition, model);
-
         if (definition.properties) {
+            model.export = 'interface';
+            model.type = 'any';
+            model.base = 'any';
+            model.default = getModelDefault(definition, model);
+
             const modelProperties = getModelProperties(openApi, definition, getModel, model);
             modelProperties.forEach(modelProperty => {
                 model.imports.push(...modelProperty.imports);
@@ -160,8 +165,18 @@ export const getModel = (
                     model.enums.push(modelProperty);
                 }
             });
+            return model;
+        } else {
+            const additionalProperties = getModel(openApi, {});
+            model.export = 'dictionary';
+            model.type = additionalProperties.type;
+            model.base = additionalProperties.base;
+            model.template = additionalProperties.template;
+            model.link = additionalProperties;
+            model.imports.push(...additionalProperties.imports);
+            model.default = getModelDefault(definition, model);
+            return model;
         }
-        return model;
     }
 
     // If the schema has a type than it can be a basic or generic type.