+++ src/openApi/v3/parser/getModels.ts
@@ -1,4 +1,5 @@
 import type { Model } from '../../../client/interfaces/Model';
+import { reservedWords } from '../../../utils/reservedWords';
 import type { OpenApi } from '../interfaces/OpenApi';
 import { getModel } from './getModel';
 import { getType } from './getType';
@@ -10,8 +11,21 @@ export const getModels = (openApi: OpenApi): Model[] => {
             if (openApi.components.schemas.hasOwnProperty(definitionName)) {
                 const definition = openApi.components.schemas[definitionName];
                 const definitionType = getType(definitionName);
-                const model = getModel(openApi, definition, true, definitionType.base);
+                const model = getModel(openApi, definition, true, definitionType.base.replace(reservedWords, '_$1'));
                 models.push(model);
+            }
+        }
+        for (const definitionName in openApi.components.parameters) {
+            if (openApi.components.parameters.hasOwnProperty(definitionName)) {
+                const definition = openApi.components.parameters[definitionName];
+                const definitionType = getType(definitionName);
+                const schema = definition.schema;
+                if (schema) {
+                    const model = getModel(openApi, schema, true, definitionType.base.replace(reservedWords, '_$1'));
+                    model.description = definition.description || null;
+                    model.deprecated = definition.deprecated;
+                    models.push(model);
+                }
             }
         }
     }