+++ src/utils/fileSystem.ts
@@ -1,50 +1,16 @@
 import {
     copyFile as __copyFile,
-    exists as __exists,
-    mkdir as __mkdir,
+    mkdirp as __mkdirp,
+    pathExists as __pathExists,
     readFile as __readFile,
-    rm as __rm,
+    remove as __remove,
     writeFile as __writeFile,
-} from 'fs';
-import { promisify } from 'util';
+} from 'fs-extra';
 
-// Wrapped file system calls
-export const readFile = promisify(__readFile);
-export const writeFile = promisify(__writeFile);
-export const copyFile = promisify(__copyFile);
-export const exists = promisify(__exists);
-
-export const mkdir = (path: string): Promise<void> =>
-    new Promise((resolve, reject) => {
-        __mkdir(
-            path,
-            {
-                recursive: true,
-            },
-            error => {
-                if (error) {
-                    reject(error);
-                } else {
-                    resolve();
-                }
-            }
-        );
-    });
-
-export const rmdir = (path: string): Promise<void> =>
-    new Promise((resolve, reject) => {
-        __rm(
-            path,
-            {
-                recursive: true,
-                force: true,
-            },
-            error => {
-                if (error) {
-                    reject(error);
-                } else {
-                    resolve();
-                }
-            }
-        );
-    });
+// Export calls (needed for mocking)
+export const readFile = __readFile;
+export const writeFile = __writeFile;
+export const copyFile = __copyFile;
+export const exists = __pathExists;
+export const mkdir = __mkdirp;
+export const rmdir = __remove;