+++ src/utils/getPattern.spec.ts
@@ -10,5 +10,7 @@ describe('getPattern', () => {
         expect(getPattern('\\')).toEqual('\\\\');
         expect(getPattern('\\/')).toEqual('\\\\/');
         expect(getPattern('\\/\\/')).toEqual('\\\\/\\\\/');
+        // eslint-disable-next-line prettier/prettier
+        expect(getPattern("'")).toEqual("\\'");
     });
 });