+++ src/utils/writeClientSchemas.spec.ts
@@ -1,3 +1,6 @@
+import { EOL } from 'os';
+import { resolve } from 'path';
+
 import type { Model } from '../client/interfaces/Model';
 import { HttpClient } from '../HttpClient';
 import { Indent } from '../Indent';
@@ -51,6 +54,6 @@ describe('writeClientSchemas', () => {
 
         await writeClientSchemas(models, templates, '/', HttpClient.FETCH, false, Indent.SPACE_4);
 
-        expect(writeFile).toBeCalledWith('/$User.ts', 'schema');
+        expect(writeFile).toBeCalledWith(resolve('/', '/$User.ts'), `schema${EOL}`);
     });
 });