+++ test/e2e/v3.fetch.spec.ts
@@ -89,7 +89,7 @@ describe('v3.fetch', () => {
         } catch (e) {
             error = (e as Error).message;
         }
-        expect(error).toContain('CancelError: Request aborted');
+        expect(error).toContain('Request aborted');
     });
 
     it('should throw known error (500)', async () => {
@@ -147,7 +147,8 @@ describe('v3.fetch', () => {
         expect(error).toBe(
             JSON.stringify({
                 name: 'ApiError',
-                message: 'Generic Error',
+                message:
+                    'Generic Error: status: 409; status text: Conflict; body: {\n  "status": 409,\n  "message": "hello world"\n}',
                 url: 'http://localhost:3000/base/api/v1.0/error?status=409',
                 status: 409,
                 statusText: 'Conflict',